requirements.txt
streamlit
rdkit-pypi
pandas
numpy
pillow
plotly
requests

